Update main.rs - change Esperanto test to pandiacritical sentence
The test as stands is not useful for testing the ability to show Esperanto text. This commit changes it to instead read for Esperanto «Echoes change every thursday» («Eĥoj ŝanĝas ĉiun ĵaŭdon»), which is nonsense, but contains every diacritc in the Esperanto alphabet, unlike the sentence about fox jumping.